requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

# Async Database
asyncpg>=0.30.0        # PostgreSQL (Supabase) — 0.30+ supports Python 3.13
aiosqlite>=0.20.0      # SQLite (local dev fallback)
sqlalchemy[asyncio]>=2.0.35
alembic>=1.13.2

# HTTP Client
httpx==0.27.2

# Auth
pyjwt==2.9.0
passlib[bcrypt]==1.7.4

# OpenAI
openai==1.51.0

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# CORS
starlette==0.38.6

# Logging
python-json-logger==2.0.7
